blog-application/requirements/base.txt
asgiref==3.5.2
Django==4.1
sqlparse==0.4.2
django-environ==0.9.0
django-model-utils==4.2.0
django-taggit==3.0.0
